HDPI-2978: Build Case Summary tab
